Began attempting an infer reader
The plan was to read the next token which could be anything, a primitive
or symbol. Then try to work out what it is, like does it match a number
regex. If not then just make a symbol.

I'm heading towards code repetition with the escaping though. Maybe
this approach will work, I just need to improve escaping. If not, back
to the drawing board.